\name{boxPlate}
\alias{boxPlate}
\title{Generate a boxplot of expression intensities for each well on the plate...}
\usage{boxPlate(eset, filePrefix="boxPlate")
}
\description{Generate a boxplot of expression intensities for each well on the plate}
\note{the color codes correspond to the sampleColor column of the pData of the ExpressionSet passed to argument
'eset'; if no sampleColor column is present, the boxplots will use 'blue' as the fill color}
\value{for each of the plates contained in the ExpressionSet object 'eset' a plot is generated}
\arguments{\item{eset}{ExpressionSet as generated by the 'preprocessing' function}
\item{filePrefix}{prefix that will be used to generate the output files; for plate <i>, the default
name will be <filePrefix>Plate<i>.pdf; the default file prefix is "boxPlate"}
}

