\name{densityPlate}
\alias{densityPlate}
\title{Generate for each well on the plate a density plot of the normalized probeset values...}
\usage{densityPlate(eset, filePrefix="densityPlate", title)
}
\description{Generate for each well on the plate a density plot of the normalized probeset values}
\details{The grey densities displayed on each well plot
correspond to the individual densities of all other wells; the colored density
is the density of the given well (plate layout is respected).}
\note{the color codes correspond to the sampleColor column of the pData of the ExpressionSet passed to argument
'eset'; if no sampleColor column is present, the density for a given sample will be plotted in 'blue'}
\value{no return value; a graph is drawn to the current device}
\arguments{\item{eset}{ExpressionSet as generated by the 'preprocessing' function}
\item{filePrefix}{prefix that will be used to generate the output files; for plate <i>, the default
name will be <filePrefix>Plate<i>.png; the default file prefix is "densityPlate"}
\item{title}{title name of the variable in the pData of the eset ExpresionSet that can be used to provide titles
on the individual well plots; if NULL no titles are plotted for the individual wells}
}

