\name{plot.ahullS}
\alias{plot.ahullS}
\title{Patched version of the plot...}
\usage{plot.ahullS(x, add=FALSE, do.shape=FALSE, wlines=c("none", "both", "del", "vor"),
    wpoints=TRUE, number=FALSE, col, xlim, ylim, lwd, ...)
}
\description{Patched version of the plot.ahull function from the alphahull package
This version TODO}
\value{plot is displayed on the current device}
\arguments{\item{x}{x}
\item{add}{add}
\item{do.shape}{do.shape}
\item{wlines}{wlines}
\item{wpoints}{wpoints}
\item{number}{number}
\item{col}{col}
\item{xlim}{xlim}
\item{ylim}{ylim}
\item{lwd}{lwd}
\item{...}{further arguments}
}

