\name{getPosition}
\alias{getPosition}
\title{Get the Position on the Plate for One Single CEL File...}
\usage{getPosition(fileName)
}
\description{Get the Position on the Plate for One Single CEL File}
\value{data frame with columns titanPlateBarcode, titanRow, titanColumn}
\arguments{\item{fileName}{path to CEL file}
}

