\name{statisticPlate}
\alias{statisticPlate}
\title{Adds a statistic to the pData of the ExpressionSet object that indicates certain types...}
\usage{statisticPlate(eset)
}
\description{Adds a statistic to the pData of the ExpressionSet object that indicates certain types
of deviations in the MA plots}
\value{object of class ExpressionSet}
\note{this function is experimental and currently only a research idea}
\arguments{\item{eset}{ExpressionSet as generated by the 'preprocessing' function}
}

